lambda/requirements.txt
fastapi==0.95.2
mangum==0.17.0
boto3==1.34.11
opensearch-py==2.4.2
requests-aws4auth==1.2.3
python-multipart==0.0.6
pydantic==1.10.13
pydantic-core==2.4.0
typing-extensions==4.9.0
pillow==10.4.0

